requirements.txt
frida-tools
pyinstaller
